Add EditHelper
